Fix db handling with prefork daemons
So far the code did db connect before a preforking
server created the forks. This led to multiple
processes using the same db connection and mixing
up transactions (or better say, generally db access).

The solution is to hide the creation of the db
handle, which in our case is inside the balib
object as an accessor with a default sub which
creates that object lazily on first access.

Additionally you MUST NOT call that accessor
inside startup() because that would initialize
it before the prefork.

This patch is the refactoring according to that.